refactor: src 폴더 추가 및 셋팅 수정
