+++ .eslintrc
@@ -1,0 +1,28 @@
+{
+	"env": {
+		"browser": true,
+		"es2021": true,
+		"node": true
+	},
+	"parser": "@typescript-eslint/parser",
+	"parserOptions": {
+		"project": "./tsconfig.json"
+	},
+	"plugins": ["@typescript-eslint"],
+	"extends": [
+		"airbnb",
+		"airbnb-typescript",
+		"airbnb/hooks",
+		"next",
+		"next/core-web-vitals",
+		"plugin:prettier/recommended"
+	],
+	"rules": {
+		"@next/next/no-document-import-in-page": "off",
+		"@typescript-eslint/no-shadow": "off",
+		"@typescript-eslint/return-await": "off",
+		"import/no-named-as-default": "off",
+		"import/prefer-default-export": "off",
+		"react/jsx-props-no-spreading": "off"
+	}
+}
