+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,24 +1,26 @@
-# 📣 작업 내용 힌줄 요약
+# PR
+
+## 📣 작업 내용 힌줄 요약
 
 <!-- 작업한 내용에 대한 간략한 내용을 적어주세요 -->
 
-# ✍🏼 작업 내용 상세 설명
+## ✍🏼 작업 내용 상세 설명
 
 <!-- 작업한 내용에 대한 자세한 설명을 적어주세요 -->
 
-# 🌆 스크린샷
+## 🌆 스크린샷
 
 <!-- 작업한 내용에 대한 스크린샷이 있다면 추가해주세요 -->
 
-# ❓ 기타 질문 및 특이사항
+## ❓ 기타 질문 및 특이사항
 
 <!-- 개발 중 질문이 있다면 남겨주세요 -->
 
-# 🔗 레퍼런스
+## 🔗 레퍼런스
 
 <!-- 참고할 수 있는 링크가 있다면 공유해주세요 -->
 
-# ✅ 체크리스트
+## ✅ 체크리스트
 
 <!-- 아래 사항들이 전부 체크되어야만 review 및 merge 됩니다 -->
