--- components/global/Footer.tsx
@@ -1,5 +1,0 @@
-const Footer = (): JSX.Element => {
-  return <footer>footer</footer>;
-};
-
-export default Footer;
