--- components/global/HeadInfo.tsx
@@ -1,19 +1,0 @@
-import Head from 'next/head';
-
-interface IHeadInfo {
-  title: string;
-  keywords: string;
-  description: string;
-}
-
-const HeadInfo = ({ title, keywords, description }: IHeadInfo): JSX.Element => {
-  return (
-    <Head>
-      <title>{title}</title>
-      <meta name="keywords" content={keywords} />
-      <meta name="description" content={description} />
-    </Head>
-  );
-};
-
-export default HeadInfo;
