--- components/global/Navbar.tsx
@@ -1,5 +1,0 @@
-const Navbar = (): JSX.Element => {
-  return <nav>nav bar</nav>;
-};
-
-export default Navbar;
