--- library/redux/modules/testSlice.ts
@@ -1,24 +1,0 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-interface ITestState {
-  value: string[];
-}
-
-const initialState: ITestState = {
-  value: [],
-};
-
-export const testSlice = createSlice({
-  name: 'test',
-  initialState,
-  reducers: {
-    //actions here
-    addTest: (state, action: PayloadAction<string>) => {
-      state.value.push(action.payload);
-    },
-  },
-});
-
-export const { addTest } = testSlice.actions;
-
-export default testSlice.reducer;
