+++ package.json
@@ -9,44 +9,41 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "11.1.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-query": "^3.25.1",
-    "react-redux": "^7.2.5",
-    "redux-typescript": "^1.2.1",
     "@emotion/css": "^11.1.3",
     "@emotion/react": "^11.4.1",
     "@emotion/server": "^11.4.0",
     "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/styles": "^4.10.0",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/styles": "^4.11.4",
     "@mui/material": "^5.0.3",
     "@reduxjs/toolkit": "^1.6.2",
-    "@types/material-ui": "^0.21.8",
-    "@types/react-redux": "^7.1.19",
-    "axios": "^0.22.0"
+    "axios": "^0.22.0",
+    "next": "11.1.2",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-query": "^3.26.0",
+    "react-redux": "^7.2.5",
+    "redux-typescript": "^1.2.1"
   },
   "devDependencies": {
+    "@types/material-ui": "^0.21.9",
     "@types/node": "^16.10.3",
     "@types/react": "^17.0.27",
     "@types/react-dom": "^17.0.9",
+    "@types/react-redux": "^7.1.19",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
-    "eslint": "^7.32.0",
+    "eslint": "7.32.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-config-airbnb-typescript": "^14.0.1",
     "eslint-config-next": "11.1.2",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "prettier": "^2.4.1",
-    "ts-node": "^10.2.1",
-    "typescript": "4.4.3",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-import-resolver-babel-module": "^5.3.1",
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react-hooks": "^4.2.0"
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.4.1",
+    "ts-node": "^10.2.1",
+    "typescript": "4.4.3"
   }
 }
