--- pages/applicant-list/index.tsx
@@ -1,5 +1,0 @@
-const index = (): JSX.Element => {
-  return <div>Enter</div>;
-};
-
-export default index;
