--- pages/applicantform/index.tsx
@@ -1,5 +1,0 @@
-const ApplicantForm = (): JSX.Element => {
-  return <div>지원자정보 페이지/ 지원하기 페이지</div>;
-};
-
-export default ApplicantForm;
