--- pages/apply-history/index.tsx
@@ -1,5 +1,0 @@
-const MyHistory = (): JSX.Element => {
-  return <div>나의 지원 내역</div>;
-};
-
-export default MyHistory;
