--- pages/detail/detail.tsx
@@ -1,5 +1,0 @@
-const Detail = (): JSX.Element => {
-  return <div>성계님 작업 여기</div>;
-};
-
-export default Detail;
