--- pages/index.tsx
@@ -1,7 +1,0 @@
-import type { NextPage } from 'next';
-
-const Home: NextPage = (): JSX.Element => {
-  return <div>main</div>;
-};
-
-export default Home;
