--- pages/login/index.tsx
@@ -1,5 +1,0 @@
-const Login = (): JSX.Element => {
-  return <div>로그인</div>;
-};
-
-export default Login;
