--- pages/myinfo/index.tsx
@@ -1,5 +1,0 @@
-const MyInfo = (): JSX.Element => {
-  return <div>내정보</div>;
-};
-
-export default MyInfo;
