--- pages/post-history/index.tsx
@@ -1,5 +1,0 @@
-const PostHistory = (): JSX.Element => {
-  return <div>공고 내역</div>;
-};
-
-export default PostHistory;
