--- pages/post-register/index.tsx
@@ -1,5 +1,0 @@
-const PostRegister = (): JSX.Element => {
-  return <div>Enter</div>;
-};
-
-export default PostRegister;
