--- pages/signup/index.tsx
@@ -1,5 +1,0 @@
-const Signup = (): JSX.Element => {
-  return <div>회원가입</div>;
-};
-
-export default Signup;
