+++ src/components/common/Footer.tsx
@@ -1,0 +1,5 @@
+const Footer: React.FC = () => {
+  return <footer>footer</footer>;
+};
+
+export default Footer;
