+++ src/components/common/Navbar.tsx
@@ -1,0 +1,5 @@
+const Navbar: React.FC = () => {
+  return <nav>nav bar</nav>;
+};
+
+export default Navbar;
