+++ src/pages/_app.tsx
@@ -1,0 +1,46 @@
+import { useEffect } from 'react';
+import CssBaseline from '@mui/material/CssBaseline';
+import { ThemeProvider } from '@mui/material/styles';
+import type { AppProps } from 'next/app';
+import { Provider } from 'react-redux';
+import HeadInfo from 'src/components/common/HeadInfo';
+import Layout from 'src/components/common/Layout';
+import { useWindowSize } from 'src/library/hooks/useWindowSize';
+import store from 'src/library/redux';
+import theme from 'src/styles/theme';
+
+const MyApp: React.FC<AppProps> = ({ Component, pageProps }) => {
+  const { windowSize } = useWindowSize();
+
+  useEffect(() => {
+    let vh = window.innerHeight * 0.01;
+    document.documentElement.style.setProperty('--vh', `${vh}px`);
+  }, [windowSize.height]);
+
+  useEffect(() => {
+    // Remove the server-side injected CSS.
+    const jssStyles = document.querySelector('#jss-server-side');
+    if (jssStyles) {
+      jssStyles?.parentElement?.removeChild(jssStyles);
+    }
+  }, []);
+
+  return (
+    <>
+      <HeadInfo
+        title="알구"
+        keywords="아르바이트, 알바, 소일거리, 중장년"
+        description="진짜 시니어를 위한 일자리"
+      />
+      <Provider store={store}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          <Layout>
+            <Component {...pageProps} />
+          </Layout>
+        </ThemeProvider>
+      </Provider>
+    </>
+  );
+};
+export default MyApp;
