+++ src/pages/applicant-list/index.tsx
@@ -1,0 +1,5 @@
+const index: React.FC = () => {
+  return <div>Enter</div>;
+};
+
+export default index;
