+++ src/pages/applicantform/index.tsx
@@ -1,0 +1,5 @@
+const ApplicantForm: React.FC = () => {
+  return <div>지원자정보 페이지/ 지원하기 페이지</div>;
+};
+
+export default ApplicantForm;
