+++ src/pages/apply-history/index.tsx
@@ -1,0 +1,5 @@
+const MyHistory: React.FC = () => {
+  return <div>나의 지원 내역</div>;
+};
+
+export default MyHistory;
