+++ src/pages/detail/detail.tsx
@@ -1,0 +1,5 @@
+const Detail: React.FC = () => {
+  return <div>성계님 작업 여기</div>;
+};
+
+export default Detail;
