+++ src/pages/index.tsx
@@ -1,0 +1,7 @@
+import type { NextPage } from 'next';
+
+const Home: NextPage = () => {
+  return <div>main</div>;
+};
+
+export default Home;
