+++ src/pages/login/index.tsx
@@ -1,0 +1,5 @@
+const MyFavorite: React.FC = () => {
+  return <div>관심목록</div>;
+};
+
+export default MyFavorite;
