+++ src/pages/myfavorite/index.tsx
