+++ src/pages/myinfo/index.tsx
@@ -1,0 +1,5 @@
+const MyInfo: React.FC = () => {
+  return <div>내 정보</div>;
+};
+
+export default MyInfo;
