+++ src/pages/post-history/index.tsx
@@ -1,0 +1,5 @@
+const PostHistory: React.FC = () => {
+  return <div>공고 내역</div>;
+};
+
+export default PostHistory;
