+++ src/pages/post-register/index.tsx
@@ -1,0 +1,5 @@
+const PostRegister: React.FC = () => {
+  return <div>Enter</div>;
+};
+
+export default PostRegister;
