+++ src/pages/signup/index.tsx
@@ -1,0 +1,5 @@
+const Signup: React.FC = () => {
+  return <div>회원가입</div>;
+};
+
+export default Signup;
